api_gateway/requirements.txt
Django==4.2.6
djangorestframework==3.15.1
python_dateutil==2.9.0.post0
